Prevent k8s from wiping out NFS mounts
we had encountered an issue where a pod would get into a strange state and the nfs volume would be wiped clean, and any subsequent files added to the volume will be wiped out within a second of being added/generated.

This fix puts some protection around the os.removeall that was responsible. The protection will unmount anything in the subdir before recursively removing the directory.